api/util/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# brayanlondono_/script.py: 3
psycopg2_binary == 2.9.1
